backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
python-dotenv==1.1.1
python-multipart==0.0.20